Refector storage class so that I have diagram to draw for DG
