--- src/main/java/seedu/duke/storage/FileStorage.java
@@ -1,4 +1,0 @@
-package seedu.duke.storage;
-
-public class FileStorage {
-}
