+++ src/main/java/seedu/duke/storage/StorageManager.java
@@ -1,37 +1,38 @@
 package seedu.duke.storage;
 
-
-import seedu.duke.models.schema.*;
-import seedu.duke.utils.Parser;
+import seedu.duke.models.schema.Event;
+import seedu.duke.models.schema.ModuleList;
+import seedu.duke.models.schema.ModuleWeekly;
+import seedu.duke.models.schema.Schedule;
+import seedu.duke.models.schema.Student;
+import seedu.duke.models.schema.TimetableUserCommand;
 import seedu.duke.utils.exceptions.CorruptedFileException;
 import seedu.duke.utils.exceptions.InvalidTimetableUserCommandException;
 import seedu.duke.utils.exceptions.MissingFileException;
 import seedu.duke.utils.exceptions.TimetableUnavailableException;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
 import java.io.IOException;
 import java.io.FileWriter;
 import java.io.BufferedWriter;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
+
+import static seedu.duke.storage.FileDecoder.createNewStorageFile;
+import static seedu.duke.storage.FileDecoder.retrieveTimetable;
+import static seedu.duke.storage.FileDecoder.retrieveStudentDetails;
+import static seedu.duke.storage.FileDecoder.retrieveSchedule;
+import static seedu.duke.storage.ResourceStorage.determineRequirements;
 
 
 public class StorageManager {
 
-    private String userDirectory = System.getProperty("user.dir");
-
+    private final String userDirectory;
 
     /**
      * Constructs a new Storage instance with the specified file path.
      */
     public StorageManager() {
-
+        this.userDirectory = System.getProperty("user.dir");
     }
 
     //@@author ryanlohyr
@@ -47,52 +48,15 @@ public class StorageManager {
         return new ArrayList<>(Arrays.asList(courseArray));
     }
 
-    /**
-     * Determines the course requirements based on the specified major. Function is used if file is not found
-     *
-     * @param major A string representing the major (e.g., "CEG" for Computer Engineering, "CS" for Computer Science).
-     * @return An array of strings containing the course requirements for the specified major.
-     */
-    private static String[] determineRequirements(String major) {
-        String[] courseArray;
 
-        String[] csCourseArray = {
-            "CS1101S", "ES2660", "GEC1000", "GEA1000", "GESS1000",
-            "GEN2000", "IS1108", "CS1231S", "CS2030", "CS2040S",
-            "CS2100", "CS2101", "CS2103T", "CS2106", "CS2109S",
-            "CS3230", "MA1521", "MA1522", "ST2334", "CP3880"
-        };
-        String[] cegCourseArray = {
-            "CG1111A", "MA1511", "MA1512", "CS1010", "GESS1000",
-            "GEC1000", "GEN2000", "ES2631", "GEA1000", "DTK1234",
-            "EG1311", "IE2141", "EE2211", "EG2501", "CDE2000",
-            "PF1101", "CG4002", "MA1508E", "EG2401A", "CP3880",
-            "CG2111A", "CS1231", "CG2023", "CG2027", "CG2028",
-            "CG2271", "ST2334", "CS2040C", "CS2113", "EE2026", "EE4204"
-        };
-
-        if(major.equals("CEG")){
-            courseArray = cegCourseArray;
-        }else{
-            courseArray = csCourseArray;
-        }
-        return courseArray;
-    }
 
     //@@author SebasFok
     /**
-     * Creates a "schedule.txt" file in the data directory.
-     *
+     * Creates a data directory containing txt folders to store student details, schedule and timetable.
      */
     public void createUserStorageFile() {
         String dataDirectory = userDirectory + "/data";
-
-        createDirectory(dataDirectory);
-
-        createFileInDirectory(dataDirectory, "schedule.txt");
-        createFileInDirectory(dataDirectory, "studentDetails.txt");
-        createFileInDirectory(dataDirectory, "timetable.txt");
-
+        createNewStorageFile(dataDirectory);
     }
 
     /**
@@ -104,69 +68,8 @@ public class StorageManager {
      * @throws CorruptedFileException  If the file is corrupted or has unexpected content.
      */
     public Schedule loadSchedule() throws MissingFileException, CorruptedFileException {
-
         String scheduleFilePath = userDirectory + "/data/schedule.txt";
-
-        if (!isFileExist(scheduleFilePath)) {
-            throw new MissingFileException();
-        }
-
-        Schedule schedule = new Schedule();
-
-        try {
-            // Create a FileReader and BufferedReader to read the file.
-            FileReader fileReader = new FileReader(scheduleFilePath);
-            BufferedReader bufferedReader = new BufferedReader(fileReader);
-
-            String line;
-
-            int targetIndex = 0;
-            int[] modulesPerSemArray = new int[]{0, 0, 0, 0, 0, 0, 0, 0};
-
-            // Read lines from the file and add them to the ArrayList.
-            while ((line = bufferedReader.readLine()) != null) {
-
-                String[] splitParts = line.split(" \\| ");
-
-                switch (splitParts[0]) {
-
-                // Happens once on the first line of txt file so that sorting subsequent modules is possible
-                case "ModulesPerSem":
-                    String[] modulesPerSemStringArray = splitParts[1].split(",");
-                    for (int i = 0; i < modulesPerSemArray.length; i++) {
-                        modulesPerSemArray[i] = Integer.parseInt(modulesPerSemStringArray[i]);
-                    }
-                    break;
-                case "Module":
-                    String module = splitParts[1];
-                    int targetSemester = 1;
-                    int indexOfLastModuleOfSem = modulesPerSemArray[targetSemester - 1] - 1;
-                    while (targetIndex > indexOfLastModuleOfSem) {
-                        indexOfLastModuleOfSem += modulesPerSemArray[targetSemester];
-                        targetSemester += 1;
-                    }
-
-                    schedule.addModule(module, targetSemester);
-                    if (splitParts[2].equals("O")) {
-                        schedule.getModule(module).markModuleAsCompleted();
-                    }
-                    targetIndex += 1;
-                    break;
-                default:
-                    if (!splitParts[0].trim().isEmpty()) {
-                        throw new CorruptedFileException();
-                    }
-                }
-            }
-
-            // Close the BufferedReader to release resources.
-            bufferedReader.close();
-        } catch (Exception e) {
-            throw new CorruptedFileException();
-        }
-
-        return schedule;
-
+        return retrieveSchedule(scheduleFilePath);
     }
 
     //@@author SebasFok
@@ -181,77 +84,7 @@ public class StorageManager {
 
         String studentDetailsFilePath = userDirectory + "/data/studentDetails.txt";
 
-        if (!isFileExist(studentDetailsFilePath)) {
-            throw new MissingFileException();
-        }
-
-        try {
-            // Create a FileReader and BufferedReader to read the file.
-            FileReader fileReader = new FileReader(studentDetailsFilePath);
-            BufferedReader bufferedReader = new BufferedReader(fileReader);
-
-            ArrayList<String> studentDetails = new ArrayList<>(3);
-
-            String line;
-            int lineNumber = 0;
-
-            // to track which line it is supposed to be on
-            HashMap<String, Integer> variableMap = new HashMap<>();
-            // Adding key-value pairs
-            variableMap.put("Name", 0);
-            variableMap.put("Major", 1);
-            variableMap.put("Year", 2);
-
-            // Read lines from the file and add them to the ArrayList.
-            while ((line = bufferedReader.readLine()) != null) {
-
-                String[] splitParts = line.split(" \\| ");
-
-                String userAttribute = splitParts[0];
-
-                //validation to see that variables has not been tampered with
-                if(variableMap.get(userAttribute) != lineNumber){
-                    throw new CorruptedFileException();
-                }
-
-                switch (splitParts[0]) {
-
-                case "Name":
-                    String name = splitParts[1];
-                    studentDetails.add(0, name);
-                    break;
-                case "Major":
-                    String major = splitParts[1];
-
-                    // Check if major stored in txt file is valid
-                    Major.valueOf(major.toUpperCase());
-
-                    studentDetails.add(1, major);
-                    break;
-                case "Year":
-                    String year = splitParts[1];
-
-                    //Check if year stored in txt file is valid
-                    if (!Parser.isValidAcademicYear(year)){
-                        throw new CorruptedFileException();
-                    }
-
-                    studentDetails.add(2, year);
-                    break;
-                default:
-                    if (!splitParts[0].trim().isEmpty()) {
-                        throw new CorruptedFileException();
-                    }
-                }
-                lineNumber += 1;
-            }
-            // Close the BufferedReader to release resources.
-            bufferedReader.close();
-
-            return studentDetails;
-        } catch (Exception e) {
-            throw new CorruptedFileException();
-        }
+        return retrieveStudentDetails(studentDetailsFilePath);
     }
 
     //@@author janelleenqi
@@ -269,44 +102,7 @@ public class StorageManager {
 
         String timetableFilePath = userDirectory + "/data/timetable.txt";
 
-        if (!isFileExist(timetableFilePath)) {
-            throw new MissingFileException();
-        }
-
-        ArrayList<TimetableUserCommand> timetableUserCommands;
-        try {
-            // Create a FileReader and BufferedReader to read the file.
-            FileReader fileReader = new FileReader(timetableFilePath);
-            BufferedReader bufferedReader = new BufferedReader(fileReader);
-
-            String line;
-            if ((line = bufferedReader.readLine()) != null) {
-                if (!line.equals("TimetableForCurrentSem")) {
-                    throw new CorruptedFileException();
-                }
-            }
-            timetableUserCommands = new ArrayList<>();
-
-            // Read lines from the file and add them to the ArrayList.
-            while ((line = bufferedReader.readLine()) != null) {
-                try {
-                    timetableUserCommands.add(new TimetableUserCommand(student,
-                            student.getTimetable().getCurrentSemesterModulesWeekly(), line));
-                } catch (InvalidTimetableUserCommandException e) {
-                    //corrupted
-                    throw new CorruptedFileException();
-                }
-
-            }
-
-            // Close the BufferedReader to release resources.
-            bufferedReader.close();
-        } catch (Exception e) {
-            throw new CorruptedFileException();
-        }
-
-        return timetableUserCommands;
-
+        return retrieveTimetable(student,timetableFilePath);
     }
 
     //@@author janelleenqi
@@ -331,8 +127,13 @@ public class StorageManager {
         }
     }
 
-    //@@author
-
+    //@@author SebasFok
+    /**
+     * Saves the student's details (name, major, and year) to the "studentDetails.txt" file.
+     *
+     * @param student The Student object containing the details to be saved.
+     * @throws IOException If an I/O error occurs while writing to the file.
+     */
     public void saveStudentDetails (Student student) throws IOException {
 
         String studentDetailsFilePath = userDirectory + "/data/studentDetails.txt";
@@ -359,7 +160,7 @@ public class StorageManager {
 
     //@@author SebasFok
     /**
-     * Saves the student's details (name, major, and year) to the "studentDetails.txt" file.
+     * Saves the student's schedule to the "schedule.txt" file.
      *
      * @param student The Student object containing the details to be saved.
      * @throws IOException If an I/O error occurs while writing to the file.
@@ -421,62 +222,6 @@ public class StorageManager {
     }
 
     // Below this comment are standard file methods
-
-    //@@author SebasFok
-    /**
-     * Takes in the location of the file in question and returns whether the file exist
-     *
-     * @param filePath location of the file in question
-     * @return return true if the file exist,return false otherwise
-     */
-    public static boolean isFileExist(String filePath) {
-        Path path = Paths.get(filePath);
-        return Files.exists(path);
-    }
-
-    //@@author SebasFok
-    /**
-     * This method takes in a path and creates a directory at that location. Should the
-     * directory already exist, no new directory will be created.
-     *
-     * @param folderPath the location of where the directory should be created
-     */
-    public static void createDirectory(String folderPath) {
-
-        File folder = new File(folderPath);
-        if (folder.mkdir()) {
-            //System.out.println("Folder created successfully.");
-        } else {
-            //System.out.println("Folder already exists");
-        }
-    }
-
-    //@@author SebasFok
-    /**
-     * This method takes in the path of a directory and creates a file 'fileName' in
-     * the directory. Should the file already exist, no new file will be created.
-     *
-     * @param directoryPath the location of the directory where the file should be created
-     * @param fileName the name of the file to be created
-     */
-    public static void createFileInDirectory(String directoryPath, String fileName) {
-
-        // Create the full path to the text file
-        String filePath = directoryPath + "/" + fileName;
-
-        File file = new File(filePath);
-
-        try {
-            // Create the file
-            if (file.createNewFile()) {
-                //System.out.println("Text file created successfully at: " + filePath);
-            } else {
-                //System.out.println("File already exists.");
-            }
-        } catch (IOException e) {
-            System.out.println("An IOException occurred: " + e.getMessage());
-        }
-    }
 
     //@@author SebasFok
     /**
